feat: finished with dark mode and responsive styling
